revert: drop unrelated WSL UNC changes from self-signed HTTPS PR
